Centralize Sources Select Component
