+++ src/main/webapp/components/sources-select/index.js
@@ -1,4 +1,2 @@
-import SourcesSelectGQL, {
-  SourcesSelectComponent as SourcesSelect,
-} from './sources-select'
+import SourcesSelectGQL, { SourcesSelect } from './sources-select'
 export { SourcesSelectGQL as default, SourcesSelect }
